[CHORE] UI View Extension 수정(#8)
