+++ Havit/Havit/Global/Extension/UIApplication+Extension.swift
@@ -1,0 +1,18 @@
+//
+//  UIApplication+Extension.swift
+//  Havit
+//
+//  Created by SHIN YOON AH on 2022/01/10.
+//
+
+import UIKit
+
+extension UIApplication {
+    var keyWindow: UIWindow? {
+        return UIApplication.shared.connectedScenes
+            .filter { $0.activationState == .foregroundActive }
+            .first(where: { $0 is UIWindowScene })
+            .flatMap({ $0 as? UIWindowScene })?.windows
+            .first(where: \.isKeyWindow)
+    }
+}
