+++ Havit/Havit/Global/Extension/UIView+Extension.swift
@@ -1,5 +1,5 @@
 //
-//  UIView+.swift
+//  UIView+Extension.swift
 //  Havit
 //
 //  Created by SHIN YOON AH on 2022/01/06.
@@ -8,27 +8,25 @@
 import UIKit
 
 extension UIView {
+    @discardableResult
     func makeShadow(_ color: UIColor,
                     _ opacity: Float,
                     _ offset: CGSize,
-                    _ radius: CGFloat) {
+                    _ radius: CGFloat) -> Self {
         layer.shadowColor = color.cgColor
         layer.shadowOpacity = opacity
         layer.shadowOffset = offset
         layer.shadowRadius = radius
+        return self
     }
     
-    @discardableResult
-    func add<T: UIView>(_ subview: T, then closure: ((T) -> Void)? = nil) -> T {
+    func add<T: UIView>(_ subview: T, completionHandler closure: ((T) -> Void)? = nil) {
         addSubview(subview)
         closure?(subview)
-        return subview
     }
     
-    @discardableResult
-    func adds<T: UIView>(_ subviews: [T], then closure: (([T]) -> Void)? = nil) -> [T] {
+    func adds<T: UIView>(_ subviews: [T], completionHandler closure: (([T]) -> Void)? = nil) {
         subviews.forEach { addSubview($0) }
         closure?(subviews)
-        return subviews
     }
 }
